+++ cratedb_toolkit/cli.py
@@ -12,6 +12,7 @@ from .io.cli import cli as io_cli
 from .job.cli import cli_list_jobs
 from .query.cli import cli as query_cli
 from .shell.cli import cli as shell_cli
+from .util.setting import init_dotenv
 
 
 @click.group(cls=ClickAliasedGroup)  # type: ignore[arg-type]
@@ -20,6 +21,7 @@ from .shell.cli import cli as shell_cli
 @click.version_option()
 @click.pass_context
 def cli(ctx: click.Context, verbose: bool, debug: bool):
+    init_dotenv()
     return boot_click(ctx, verbose, debug)